apps/moderation/pyproject.toml
[project]
name = "moderation"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "pillow (>=11.1.0,<12.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "python-ffmpeg (>=2.0.12,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "minio (>=7.2.15,<8.0.0)",
    "nats-py (>=2.10.0,<3.0.0)",
    "detoxify (>=0.5.2,<0.6.0)",
    "torch (>=2.7.0,<3.0.0)",
    "torchvision (>=0.22.0,<0.23.0)",
    "torchaudio (>=2.7.0,<3.0.0)",
    "transformers (>=4.51.3,<5.0.0)",
    "ultralytics (>=8.3.114,<9.0.0)",
    "easyocr (>=1.7.2,<2.0.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "kenlm (>=0.3.0,<0.4.0)",
    "pyctcdecode (>=0.5.0,<0.6.0)",
    "onnxruntime-gpu (>=1.21.1,<2.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "onnx (>=1.17.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu126"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch"}
torchvision = {source = "pytorch"}
torchaudio = {source = "pytorch"}
